move more items from src/http to src/ldp, fix #32
