+++ src/ldp/operations/DeleteOperation.ts
@@ -4,7 +4,7 @@ import ResourceIdentifier from '../IResourceIdentifier';
 import PermissionSet from '../../permissions/PermissionSet';
 import Conditions from '../Conditions';
 import IRepresentationPreferences from '../IRepresentationPreferences';
-import ResponseDescription from '../../http/ResponseDescription';
+import ResponseDescription from '../ResponseDescription';
 
 /**
  * Performs an LDP DELETE operation.
