+++ src/ldp/operations/LdpOperation.ts
@@ -1,6 +1,6 @@
 import PermissionSet from '../../permissions/PermissionSet';
 import IOperation from './IOperation';
-import ResponseDescription from '../../http/ResponseDescription';
+import ResponseDescription from '../ResponseDescription';
 import IRepresentation from '../IRepresentation';
 import IResourceIdentifier from '../IResourceIdentifier';
 import IRepresentationPreferences from '../IRepresentationPreferences';
